move sourceType to constants folder
