+++ signal-sweep/src/signal_sweep/core/source.py
@@ -1,20 +1,8 @@
 from dataclasses import dataclass
-from enum import Enum
 
-from .handlers.base_handler import Handler
-
-
-class SourceType(Enum):
-    CSV = "csv"
-    JSON = "json"
-    TXT = "txt"
-    # PARQUET = "parquet"
-    # EXCEL = "xlsx"
-    # TSV = "tsv"
-
+from ..shared.constants import SourceType
 
 @dataclass
 class Source:
     url: str
     type: SourceType
-    handler: Handler
