+++ signal-sweep/tests/conftest.py
@@ -15,7 +15,8 @@ import httpx
 import redis.asyncio as redis
 
 from signal_sweep.core.signal_stream import StreamData, SignalStream
-from signal_sweep.core.source import Source, SourceType
+from signal_sweep.core.source import Source
+from signal_sweep.shared.constants import SourceType
 from signal_sweep.shared.utils import AsyncProcessPoolExecutor
 from signal_sweep.core.handlers.text_handler import TextHandler
 from signal_sweep.container import ApplicationContainer
@@ -160,12 +161,10 @@ def sample_sources() -> List[Source]:
         Source(
             url="https://example.com/threat-data.txt",
             type=SourceType.TXT,
-            handler=handler,
         ),
         Source(
             url="https://example.com/other-data.txt",
             type=SourceType.TXT,
-            handler=handler,
         ),
     ]
